requirements.txt
looker-sdk==0.1.3b4
fastapi==0.42.0
pydantic==0.32.2
python-multipart==0.0.5
sendgrid==6.1.0
uvicorn==0.10.4

openpyxl==3.0.0
python-docx==0.8.10
